requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==19.3.0
Automat==0.8.0
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==20.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-classy-tags==2.0.0
django-cms==3.8.0
django-cors-headers==3.7.0
django-editorjs==0.2.1
django-filer==2.0.2
django-formtools==2.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-polymorphic==3.0.0
django-sekizai==2.0.0
django-treebeard==4.5.1
djangocms-admin-style==2.0.2
djangocms-attributes-field==2.0.0
djangocms-bootstrap4==2.0.0
djangocms-file==3.0.0
djangocms-googlemap==2.0.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-style==3.0.0
djangocms-text-ckeditor==4.0.0
djangocms-video==3.0.0
djangorestframework==3.12.4
easy-thumbnails==2.7.1
ec2-hibinit-agent==1.0.0
entrypoints==0.3
filelock==3.0.12
gunicorn==20.1.0
hibagent==1.0.1
html5lib==1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
pexpect==4.6.0
Pillow==8.2.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.4
python-debian===0.1.36ubuntu1
pytz==2021.1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.0
sqlparse==0.4.1
ssh-import-id==5.10
systemd-python==234
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
Unidecode==1.1.2
urllib3==1.25.8
virtualenv==20.4.3
wadllib==1.3.3
webencodings==0.5.1
zipp==1.0.0
zope.interface==4.7.1
